refactor(sdk): split libsdk core and language bindings
