+++ Cargo.toml
@@ -34,9 +34,12 @@ members = [
     "curvine-client-core",
     "curvine-job-client",
     "curvine-unified-fs",
+    "curvine-sdk-core",
     "curvine-bench",
     "curvine-client",
     "curvine-libsdk",
+    "curvine-libsdk-java",
+    "curvine-libsdk-python",
     "curvine-lancedb",
     'curvine-tests',
     "curvine-fuse",
@@ -67,6 +70,7 @@ default-members = [
     "curvine-client-core",
     "curvine-job-client",
     "curvine-unified-fs",
+    "curvine-sdk-core",
     "curvine-bench",
     "curvine-client",
     "curvine-libsdk",
@@ -106,9 +110,12 @@ curvine-common-macros = { path = "curvine-common/macros" }
 curvine-client-core = { path = "curvine-client-core" }
 curvine-job-client = { path = "curvine-job-client" }
 curvine-unified-fs = { path = "curvine-unified-fs" }
+curvine-sdk-core = { path = "curvine-sdk-core" }
 curvine-bench = { path = "curvine-bench" }
-curvine-client = { path = "curvine-client" }
+curvine-client = { path = "curvine-client", default-features = false }
 curvine-libsdk = { path = "curvine-libsdk" }
+curvine-libsdk-java = { path = "curvine-libsdk-java" }
+curvine-libsdk-python = { path = "curvine-libsdk-python" }
 curvine-server = { path = "curvine-server" }
 curvine-lancedb = { path = "curvine-lancedb" }
 curvine-ufs = { path = "curvine-ufs" }
