+++ curvine-libsdk-java/src/lib.rs
@@ -14,17 +14,5 @@
 
 pub use curvine_sdk_core::{FilesystemConf, LibFilesystem, LibFsReader, LibFsWriter};
 
-#[cfg(feature = "rust-sdk")]
-pub use curvine_sdk_core::filesystem;
-#[cfg(feature = "rust-sdk")]
-pub use curvine_sdk_core::job;
-#[cfg(feature = "rust-sdk")]
-pub use curvine_sdk_core::lib_curvine;
-#[cfg(feature = "rust-sdk")]
-pub use curvine_sdk_core::master;
-
-#[cfg(feature = "python-sdk")]
-pub use curvine_libsdk_python as python;
-
-#[cfg(feature = "java-sdk")]
-pub use curvine_libsdk_java as java;
+pub mod java;
+pub use self::java::*;
