+++ curvine-libsdk-python/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "curvine-libsdk-python"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[lib]
+crate-type = ["rlib", "cdylib"]
+name = "curvine_libsdk_python"
+
+[features]
+default = []
+extension-module = ["pyo3/extension-module"]
+opendal-s3 = ["curvine-sdk-core/opendal-s3"]
+opendal-oss = ["curvine-sdk-core/opendal-oss"]
+opendal-gcs = ["curvine-sdk-core/opendal-gcs"]
+opendal-azblob = ["curvine-sdk-core/opendal-azblob"]
+opendal-cos = ["curvine-sdk-core/opendal-cos"]
+opendal-webhdfs = ["curvine-sdk-core/opendal-webhdfs"]
+opendal-hdfs = ["curvine-sdk-core/opendal-hdfs"]
+opendal-hdfs-native = ["curvine-sdk-core/opendal-hdfs-native"]
+oss-hdfs = ["curvine-sdk-core/oss-hdfs"]
+internal-oss-hdfs-jindo = ["curvine-sdk-core/internal-oss-hdfs-jindo"]
+
+[dependencies]
+curvine-sdk-core = { workspace = true }
+curvine-config = { workspace = true }
+curvine-error = { workspace = true }
+orpc = { workspace = true }
+bytes = { workspace = true }
+pyo3 = { version = "0.25.1", features = ["abi3-py38", "num-complex"] }
