+++ curvine-libsdk-python/pyproject.toml
@@ -3,9 +3,9 @@ requires = ["maturin>=1,<2"]
 build-backend = "maturin"
 
 [tool.maturin]
-python-source = "python"
+python-source = "../curvine-libsdk/python"
 module-name = "curvine_libsdk._native"
-features = ["python-sdk"]
+features = ["extension-module"]
 # include = [{ path = "src/python/**", format = "sdist"}]
 
 [project]
