+++ curvine-libsdk-python/src/python/python_abi.rs
@@ -2,7 +2,7 @@
 
 use crate::python::PythonFilesystem;
 use crate::{LibFsReader, LibFsWriter};
-use curvine_common::error::{ErrorKind, FsError};
+use curvine_error::{ErrorKind, FsError};
 use orpc::sys::DataSlice;
 use orpc::sys::{FFIUtils, RawVec};
 use pyo3::exceptions::*;
