+++ curvine-sdk-core/src/core/filesystem.rs
@@ -15,10 +15,10 @@
 use crate::core::Session;
 use crate::{LibFsReader, LibFsWriter};
 use bytes::BytesMut;
-use curvine_common::conf::ClusterConf;
-use curvine_common::fs::{FileSystem, Path};
-use curvine_common::state::{FreeResult, MountInfo, MountOptions, SetAttrOpts};
-use curvine_common::FsResult;
+use curvine_config::ClusterConf;
+use curvine_error::FsResult;
+use curvine_fs_api::{FileSystem, Path};
+use curvine_model::{FreeResult, MountInfo, MountOptions, SetAttrOpts};
 use orpc::runtime::RpcRuntime;
 
 pub struct LibFilesystem {
