+++ curvine-sdk-core/src/lib.rs
@@ -12,19 +12,29 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use curvine_sdk_core::{FilesystemConf, LibFilesystem, LibFsReader, LibFsWriter};
+mod core;
+
+mod lib_filesystem;
+pub use self::lib_filesystem::LibFilesystem;
+
+mod filesystem_conf;
+pub use self::filesystem_conf::FilesystemConf;
+
+mod lib_fs_writer;
+pub use self::lib_fs_writer::LibFsWriter;
+
+mod lib_fs_reader;
+pub use self::lib_fs_reader::LibFsReader;
+
+pub mod blocking_job {
+    pub use crate::core::job::{cancel_job, get_job_status, submit_load_job};
+}
 
 #[cfg(feature = "rust-sdk")]
-pub use curvine_sdk_core::filesystem;
+pub mod filesystem;
 #[cfg(feature = "rust-sdk")]
-pub use curvine_sdk_core::job;
+pub mod job;
 #[cfg(feature = "rust-sdk")]
-pub use curvine_sdk_core::lib_curvine;
+pub mod lib_curvine;
 #[cfg(feature = "rust-sdk")]
-pub use curvine_sdk_core::master;
-
-#[cfg(feature = "python-sdk")]
-pub use curvine_libsdk_python as python;
-
-#[cfg(feature = "java-sdk")]
-pub use curvine_libsdk_java as java;
+pub mod master;
